Re-organize: seperate Scenario, GeospatialScenario
